refactor: simplify BackboneCSDI;
