+++ pypots/imputation/csdi/core.py
@@ -1,6 +1,7 @@
 # Created by Wenjie Du <wenjay.du@gmail.com>
 # License: BSD-3-Clause
 
+import torch
 import torch.nn as nn
 
 from ...nn.modules.csdi import BackboneCSDI
@@ -9,10 +10,10 @@ from ...nn.modules.csdi import BackboneCSDI
 class _CSDI(nn.Module):
     def __init__(
         self,
+        n_pred_features,
         n_layers,
         n_heads,
         n_channels,
-        d_target,
         d_time_embedding,
         d_feature_embedding,
         d_diffusion_embedding,
@@ -24,11 +25,19 @@ class _CSDI(nn.Module):
     ):
         super().__init__()
 
+        self.n_pred_features = n_pred_features
+        self.d_time_embedding = d_time_embedding
+        self.is_unconditional = is_unconditional
+
+        self.embed_layer = nn.Embedding(
+            num_embeddings=n_pred_features,
+            embedding_dim=d_feature_embedding,
+        )
         self.backbone = BackboneCSDI(
             n_layers,
             n_heads,
             n_channels,
-            d_target,
+            n_pred_features,
             d_time_embedding,
             d_feature_embedding,
             d_diffusion_embedding,
@@ -39,6 +48,41 @@ class _CSDI(nn.Module):
             beta_end,
         )
 
+    @staticmethod
+    def time_embedding(pos, d_model=128):
+        pe = torch.zeros(pos.shape[0], pos.shape[1], d_model).to(pos.device)
+        position = pos.unsqueeze(2)
+        div_term = 1 / torch.pow(
+            10000.0, torch.arange(0, d_model, 2, device=pos.device) / d_model
+        )
+        pe[:, :, 0::2] = torch.sin(position * div_term)
+        pe[:, :, 1::2] = torch.cos(position * div_term)
+        return pe
+
+    def get_side_info(self, observed_tp, cond_mask):
+        B, K, L = cond_mask.shape
+        device = observed_tp.device
+        time_embed = self.time_embedding(
+            observed_tp, self.d_time_embedding
+        )  # (B,L,emb)
+        time_embed = time_embed.to(device)
+        time_embed = time_embed.unsqueeze(2).expand(-1, -1, K, -1)
+        feature_embed = self.embed_layer(
+            torch.arange(self.n_pred_features).to(device)
+        )  # (K,emb)
+        feature_embed = feature_embed.unsqueeze(0).unsqueeze(0).expand(B, L, -1, -1)
+
+        side_info = torch.cat(
+            [time_embed, feature_embed], dim=-1
+        )  # (B,L,K,emb+d_feature_embedding)
+        side_info = side_info.permute(0, 3, 2, 1)  # (B,*,K,L)
+
+        if not self.is_unconditional:
+            side_mask = cond_mask.unsqueeze(1)  # (B,1,K,L)
+            side_info = torch.cat([side_info, side_mask], dim=1)
+
+        return side_info
+
     def forward(self, inputs, training=True, n_sampling_times=1):
         results = {}
         if training:  # for training
@@ -48,7 +92,7 @@ class _CSDI(nn.Module):
                 inputs["cond_mask"],
                 inputs["observed_tp"],
             )
-            side_info = self.backbone.get_side_info(observed_tp, cond_mask)
+            side_info = self.get_side_info(observed_tp, cond_mask)
             training_loss = self.backbone.calc_loss(
                 observed_data, cond_mask, indicating_mask, side_info, training
             )
@@ -60,7 +104,7 @@ class _CSDI(nn.Module):
                 inputs["cond_mask"],
                 inputs["observed_tp"],
             )
-            side_info = self.backbone.get_side_info(observed_tp, cond_mask)
+            side_info = self.get_side_info(observed_tp, cond_mask)
             validating_loss = self.backbone.calc_loss_valid(
                 observed_data, cond_mask, indicating_mask, side_info, training
             )
@@ -71,7 +115,7 @@ class _CSDI(nn.Module):
                 inputs["cond_mask"],
                 inputs["observed_tp"],
             )
-            side_info = self.backbone.get_side_info(observed_tp, cond_mask)
+            side_info = self.get_side_info(observed_tp, cond_mask)
             samples = self.backbone(
                 observed_data, cond_mask, side_info, n_sampling_times
             )  # (n_samples, n_sampling_times, n_features, n_steps)
